Student aggregation total count display
