+++ components/common/ExpandableSearchGrid/expandableSearchGrid.tsx
@@ -4,6 +4,7 @@ import Card from '@mui/material/Card';
 import { CardContent } from '@mui/material';
 import { SearchBar } from '../SearchBar/searchBar';
 import React from 'react';
+import { useRouter } from 'next/router';
 
 type SearchQuery = {
   prefix?: string;
@@ -14,7 +15,7 @@ type SearchQuery = {
 
 type ExpandableSearchGridProps = {
   onChange: Function;
-  startingData: SearchQuery[];
+  studentTotals: number[];
 };
 
 /**
@@ -24,7 +25,7 @@ type ExpandableSearchGridProps = {
  */
 export const ExpandableSearchGrid = ({
   onChange,
-  startingData,
+  studentTotals,
 }: ExpandableSearchGridProps) => {
   const [value, setValue] = useState<SearchQuery[]>([]);
   const [searchTerms, setSearchTerms] = useState<SearchQuery[]>([]);
@@ -33,10 +34,6 @@ export const ExpandableSearchGrid = ({
   useEffect(() => {
     onChange(searchTerms);
   }, [onChange, searchTerms]);
-
-  useEffect(() => {
-    setSearchTerms(startingData);
-  }, [startingData]);
 
   function addSearchTerm(newSearchTerm: SearchQuery) {
     if (newSearchTerm != null) {
@@ -66,13 +63,21 @@ export const ExpandableSearchGrid = ({
       setSearchDisable(false);
     }
   }, [searchTerms]);
+  
+  const router = useRouter();
+  
+  useEffect(() => {
+    if (router.isReady) {
+      setSearchTerms(parseURIEncodedSearchTerms(router.query.searchTerms));
+    }
+  }, [router.isReady]);
 
   return (
     <div className="w-full min-h-[72px] grid grid-flow-row auto-cols-fr md:grid-flow-col justify-center">
       {searchTerms.map((option: SearchQuery, index: number) => (
         <SearchTermCard
           primaryText={searchQueryLabel(option)}
-          secondaryText={''}
+          secondaryText={studentTotalFormatter(studentTotals[index])}
           key={index}
           index={index}
           legendColor={colors[index]}
@@ -106,6 +111,13 @@ export const ExpandableSearchGrid = ({
   );
 };
 
+function studentTotalFormatter(total: number) {
+  if (total === -1) {
+    return 'Loading...';
+  }
+  return total + ' students';
+}
+
 function searchQueryLabel(query: SearchQuery): string {
   let result = '';
   if (query.prefix !== undefined) {
@@ -121,6 +133,77 @@ function searchQueryLabel(query: SearchQuery): string {
     result += ' ' + query.sectionNumber;
   }
   return result.trim();
+}
+
+function parseURIEncodedSearchTerms(
+  encodedSearchTerms: string | string[] | undefined,
+): SearchQuery[] {
+  if (encodedSearchTerms === undefined) {
+    return [];
+  } else if (typeof encodedSearchTerms === 'string') {
+    return encodedSearchTerms
+      .split(',')
+      .map((term) => parseURIEncodedSearchTerm(term));
+  } else {
+    return encodedSearchTerms.map((term) => parseURIEncodedSearchTerm(term));
+  }
+}
+
+function parseURIEncodedSearchTerm(encodedSearchTerm: string): SearchQuery {
+  let encodedSearchTermParts = encodedSearchTerm.split(' ');
+  // Does it start with prefix
+  if (/^([A-Z]{2,4})$/.test(encodedSearchTermParts[0])) {
+    // If it is just the prefix, return that
+    if (encodedSearchTermParts.length == 1) {
+      return { prefix: encodedSearchTermParts[0] };
+    }
+    // Is the second part a course number only
+    if (/^([0-9A-Z]{4})$/.test(encodedSearchTermParts[1])) {
+      if (encodedSearchTermParts.length == 2) {
+        return {
+          prefix: encodedSearchTermParts[0],
+          number: encodedSearchTermParts[1],
+        };
+      } else {
+        return {
+          prefix: encodedSearchTermParts[0],
+          number: encodedSearchTermParts[1],
+          professorName:
+            encodedSearchTermParts[2] + ' ' + encodedSearchTermParts[3],
+        };
+      }
+    }
+    // Is the second part a course number and section
+    else if (/^([0-9A-Z]{4}\.[0-9A-Z]{3})$/.test(encodedSearchTermParts[1])) {
+      let courseNumberAndSection: string[] =
+        encodedSearchTermParts[1].split('.');
+      if (encodedSearchTermParts.length == 2) {
+        return {
+          prefix: encodedSearchTermParts[0],
+          number: courseNumberAndSection[0],
+          sectionNumber: courseNumberAndSection[1],
+        };
+      } else {
+        return {
+          prefix: encodedSearchTermParts[0],
+          number: courseNumberAndSection[0],
+          sectionNumber: courseNumberAndSection[1],
+          professorName:
+            encodedSearchTermParts[2] + ' ' + encodedSearchTermParts[3],
+        };
+      }
+    }
+    // the second part is the start of the name
+    else {
+      return {
+        prefix: encodedSearchTermParts[0],
+        professorName:
+          encodedSearchTermParts[1] + ' ' + encodedSearchTermParts[2],
+      };
+    }
+  } else {
+    return { professorName: encodedSearchTerm.trim() };
+  }
 }
 
 const colors = ['#eb5757', '#2d9cdb', '#499F68'];
