+++ components/common/SearchTermCard/searchTermCard.tsx
@@ -47,10 +47,9 @@ export const SearchTermCard = (props: SearchTermCardProps) => {
             marginLeft: '8px',
           }}
         />
-        <Typography className="text-lg text-gray-600">
+        <Typography className="leading-tight text-lg text-gray-600">
           {props.primaryText}
-          <br />
-          <span className="text-base text-gray-600">{props.secondaryText}</span>
+          <span className="block text-sm text-gray-500">{props.secondaryText}</span>
         </Typography>
       </div>
       <div className="float-right">
